requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.1
python-multipart==0.0.6
redis==5.2.1
requests==2.32.3
opencv-python==4.11.0.86
websockets==12.0
python-socketio==5.10.0
jinja2==3.1.2
pydantic-settings==2.1.0
scikit-fuzzy==0.4.2
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0
matplotlib==3.10.3